% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_between}
\alias{calculate_lambda_between}
\alias{calculate_lambda_between,ECProb-method}
\title{Calculate Lambda Between Two Vertex Sets (Optimized Method)}
\usage{
calculate_lambda_between(object, set1, set2)

\S4method{calculate_lambda_between}{ECProb}(object, set1, set2)
}
\arguments{
\item{object}{An ECProb object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{A character vector of vertex names.}
}
\value{
The numeric lambda parameter.
}
\description{
Calculates the lambda parameter between two vertex sets using a
speed-optimized and accurate algorithm. This is the recommended method for accurate calculation.
}
\details{
This method has a time complexity of O(n*log(n) + m*log(m)).
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_between(ECProb)}: Method for ECProb objects.

}}
