% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_between_fast}
\alias{calculate_lambda_between_fast}
\alias{calculate_lambda_between_fast,ECProb-method}
\title{Calculate Lambda Between Two Vertex Sets (Fast Method)}
\usage{
calculate_lambda_between_fast(object, set1, set2)

\S4method{calculate_lambda_between_fast}{ECProb}(object, set1, set2)
}
\arguments{
\item{object}{An ECProb object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{A character vector of vertex names.}
}
\value{
The approximate numeric lambda parameter.
}
\description{
Calculates an approximation of the lambda parameter between two
vertex sets.
}
\details{
This method is the fastest, with a complexity of O(n+m).
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_between_fast(ECProb)}: Method for ECProb objects.

}}
