% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_in}
\alias{calculate_lambda_in}
\alias{calculate_lambda_in,ECProb-method}
\title{Calculate Lambda Within a Vertex Set (Optimized Method)}
\usage{
calculate_lambda_in(object, set)

\S4method{calculate_lambda_in}{ECProb}(object, set)
}
\arguments{
\item{object}{An ECProb object.}

\item{set}{A character vector of vertex names.}
}
\value{
The numeric lambda parameter.
}
\description{
Calculates the lambda parameter for edges within a single set
of vertices using a speed-optimized and accurate algorithm.
}
\details{
This method has a time complexity of O(n*log(n)).
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_in(ECProb)}: Method for ECProb objects.

}}
