% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_in_fast}
\alias{calculate_lambda_in_fast}
\alias{calculate_lambda_in_fast,ECProb-method}
\title{Calculate Lambda Within a Vertex Set (Fast Method)}
\usage{
calculate_lambda_in_fast(object, set)

\S4method{calculate_lambda_in_fast}{ECProb}(object, set)
}
\arguments{
\item{object}{An ECProb object.}

\item{set}{A character vector of vertex names.}
}
\value{
The approximate numeric lambda parameter.
}
\description{
Calculates an approximation of the lambda parameter for edges
within a single set of vertices.
}
\details{
This method is the fastest, with a complexity of O(n).
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_in_fast(ECProb)}: Method for ECProb objects.

}}
