% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_in_naive}
\alias{calculate_lambda_in_naive}
\alias{calculate_lambda_in_naive,ECProb-method}
\title{Calculate Lambda Within a Vertex Set (Naive Method)}
\usage{
calculate_lambda_in_naive(object, set)

\S4method{calculate_lambda_in_naive}{ECProb}(object, set)
}
\arguments{
\item{object}{An ECProb object.}

\item{set}{A character vector of vertex names.}
}
\value{
The numeric lambda parameter.
}
\description{
Calculates the lambda parameter for edges within a single set
of vertices using a direct, unoptimized summation.
}
\details{
The naive method has a time complexity of O(n^2).
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_in_naive(ECProb)}: Method for ECProb objects.

}}
