% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\docType{class}
\name{ECProb-class}
\alias{ECProb-class}
\alias{show,ECProb-method}
\title{ECProb S4 Class}
\usage{
\S4method{show}{ECProb}(object)
}
\arguments{
\item{object}{An ECProb object}
}
\description{
Extends the ECGraph class to include pre-calculated properties
necessary for edge-count probability calculations.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(ECProb)}: Shows ECGraph statistics plus the pre-calculated graph properties (size, order, average degree).

}}
\section{Slots}{

\describe{
\item{\code{graph_size}}{numeric. The total number of edges in the graph (M).}

\item{\code{graph_order}}{numeric. The total number of vertices in the graph (N).}

\item{\code{average_degree}}{numeric. The average degree of the vertices in the graph.}

\item{\code{adj}}{list. (Inherited from ECGraph) The adjacency list.}

\item{\code{degrees}}{list. (Inherited from ECGraph) The degree of each vertex.}

\item{\code{names}}{character. (Inherited from ECGraph) The names of all vertices.}
}}

\seealso{
\code{\link{ECProb}} for the constructor.
}
