% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_between}
\alias{get_edge_count_between}
\alias{get_edge_count_between,ECGraph-method}
\title{Get Edge Count Between Two Sets of Vertices}
\usage{
get_edge_count_between(object, set1, set2)

\S4method{get_edge_count_between}{ECGraph}(object, set1, set2)
}
\arguments{
\item{object}{An ECGraph object.}

\item{set1}{A character vector of vertex names for the first set.}

\item{set2}{A character vector of vertex names for the second set.}
}
\value{
An integer count of the edges between the two sets.
}
\description{
Calculates the number of edges connecting two specified sets of vertices.
Edges within the intersection of the sets are not counted.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_between(ECGraph)}: Method for ECGraph objects.

}}
\examples{
edge_df <- data.frame(p1 = c("A", "B", "C"), p2 = c("X", "Y", "A"))
ecg <- ECGraph(edge_df)
get_edge_count_between(ecg, c("A", "B"), c("X", "Y")) # Returns 2 (A-X, B-Y)
get_edge_count_between(ecg, c("A", "B"), c("C", "X")) # Returns 2 (A-X, A-C)

}
