% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_between_max}
\alias{get_edge_count_between_max}
\alias{get_edge_count_between_max,ECGraph-method}
\title{Get Maximum Possible Edge Count Between Two Sets}
\usage{
get_edge_count_between_max(object, set1, set2)

\S4method{get_edge_count_between_max}{ECGraph}(object, set1, set2)
}
\arguments{
\item{object}{An ECGraph object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{A character vector of vertex names.}
}
\value{
The maximum possible number of edges between the sets.
}
\description{
Calculates the maximum possible number of edges between two disjoint
sets of vertices (i.e., a complete bipartite graph). Only vertices
present in the ECGraph object are considered.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_between_max(ECGraph)}: Method for ECGraph objects.

}}
\examples{
ecg <- ECGraph(data.frame(p1=c("A","B","X","Y"), p2=c("B","A","Y","X")))
# "Z" is ignored as it's not in the graph.
# The result is length(c("A","B")) * length(c("X","Y")) = 2 * 2 = 4.
get_edge_count_between_max(ecg, c("A", "B"), c("X", "Y", "Z"))

}
