% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_between_max_fds}
\alias{get_edge_count_between_max_fds}
\alias{get_edge_count_between_max_fds,ECGraph-method}
\title{Get Max Edge Count Between Sets (Fixed Degree Sequence)}
\usage{
get_edge_count_between_max_fds(object, set1, set2)

\S4method{get_edge_count_between_max_fds}{ECGraph}(object, set1, set2)
}
\arguments{
\item{object}{An ECGraph object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{A character vector of vertex names.}
}
\value{
The maximum possible number of edges, while preserving vertex degrees.
}
\description{
Calculates the maximum possible number of edges between two disjoint
sets of vertices, given their prescribed degrees (Fixed Degree Sequence model).
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_between_max_fds(ECGraph)}: Method for ECGraph objects.

}}
\examples{
# Bipartite graph example
edge_df <- data.frame(p1=c("A","A","B"), p2=c("X","Y","Y"))
ecg <- ECGraph(edge_df)
get_edge_count_between_max_fds(ecg, c("A","B"), c("X","Y")) # Returns 3

}
