% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_in}
\alias{get_edge_count_in}
\alias{get_edge_count_in,ECGraph-method}
\title{Get Edge Count Within a Set of Vertices}
\usage{
get_edge_count_in(object, vertices)

\S4method{get_edge_count_in}{ECGraph}(object, vertices)
}
\arguments{
\item{object}{An ECGraph object.}

\item{vertices}{A character vector of vertex names.}
}
\value{
An integer count of the edges within the set.
}
\description{
Calculates the number of edges connecting vertices *within* a given set.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_in(ECGraph)}: Method for ECGraph objects.

}}
\examples{
edge_df <- data.frame(p1 = c("A", "B", "C", "D"), p2 = c("B", "C", "A", "E"))
ecg <- ECGraph(edge_df)
# Count edges within the set {A, B, C} (A-B, B-C, C-A)
get_edge_count_in(ecg, c("A", "B", "C")) # Returns 3

}
