% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_in_max}
\alias{get_edge_count_in_max}
\alias{get_edge_count_in_max,ECGraph-method}
\title{Get Maximum Possible Edge Count Within a Set}
\usage{
get_edge_count_in_max(object, vertices)

\S4method{get_edge_count_in_max}{ECGraph}(object, vertices)
}
\arguments{
\item{object}{An ECGraph object.}

\item{vertices}{A character vector of vertex names.}
}
\value{
The maximum possible number of edges.
}
\description{
Calculates the maximum possible number of edges in a simple graph
formed by a given set of vertices (i.e., a complete graph). Only vertices
present in the ECGraph object are considered.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_in_max(ECGraph)}: Method for ECGraph objects.

}}
\examples{
ecg <- ECGraph(data.frame(p1=c("A","B","C","D"), p2=c("B","C","D","A")))
# The function ignores vertices not in the graph, like "X"
# It calculates max edges for the valid set {A, B, C}, which is 3.
get_edge_count_in_max(ecg, c("A", "B", "C", "X"))

}
