% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_edge_count_in_max_fds}
\alias{get_edge_count_in_max_fds}
\alias{get_edge_count_in_max_fds,ECGraph-method}
\title{Get Max Edge Count Within a Set (Fixed Degree Sequence)}
\usage{
get_edge_count_in_max_fds(object, vertices)

\S4method{get_edge_count_in_max_fds}{ECGraph}(object, vertices)
}
\arguments{
\item{object}{An ECGraph object.}

\item{vertices}{A character vector of vertex names.}
}
\value{
The maximum possible number of edges, while preserving vertex degrees.
}
\description{
Calculates the maximum possible number of edges within a set of
vertices, given their prescribed degrees (Fixed Degree Sequence model).
}
\section{Methods (by class)}{
\itemize{
\item \code{get_edge_count_in_max_fds(ECGraph)}: Method for ECGraph objects.

}}
\examples{
# Example from a known graphic sequence
ecg <- ECGraph(data.frame(p1=c("A","B","C"), p2=c("B","C","A"))) # All degrees are 2
get_edge_count_in_max_fds(ecg, c("A", "B", "C")) # Returns 3

}
