% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{get_neighbors}
\alias{get_neighbors}
\alias{get_neighbors,ECGraph-method}
\title{Get Neighbors of Vertices}
\usage{
get_neighbors(object, vertices)

\S4method{get_neighbors}{ECGraph}(object, vertices)
}
\arguments{
\item{object}{An ECGraph object.}

\item{vertices}{A character vector of vertex names.}
}
\value{
A character vector of unique neighbor names.
}
\description{
Retrieves a vector of all unique neighbors for a given
set of vertices.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_neighbors(ECGraph)}: Method for ECGraph objects.

}}
\examples{
edge_df <- data.frame(v1 = c("A", "B", "C"), v2 = c("B", "C", "D"))
ecg <- ECGraph(edge_df)
# Get neighbors of vertices A and C
get_neighbors(ecg, c("A", "C")) # Returns "B" and "D"

}
