% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{reduce_universe_by_terms}
\alias{reduce_universe_by_terms}
\alias{reduce_universe_by_terms,ECTermScoring-method}
\title{Reduce the Universe of an ECTermScoring Object to a Set of Terms}
\usage{
reduce_universe_by_terms(object, terms_to_keep)

\S4method{reduce_universe_by_terms}{ECTermScoring}(object, terms_to_keep)
}
\arguments{
\item{object}{An ECTermScoring object.}

\item{terms_to_keep}{A character vector of term IDs to keep.}
}
\value{
A new, smaller ECTermScoring object.
}
\description{
Creates a new, smaller ECTermScoring object by restricting it
to a specified subset of terms.
}
\details{
This is a convenience method which also ensures that statistical tests are correctly
conditioned. It takes an ECTermScoring object and a vector of terms, finds
the intersection, and returns a new object where all properties (degrees, graph
size, etc.) are recalculated based on the input constraint.
}
\section{Methods (by class)}{
\itemize{
\item \code{reduce_universe_by_terms(ECTermScoring)}: Method for ECTermScoring objects.

}}
