% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{remove_empty_terms}
\alias{remove_empty_terms}
\alias{remove_empty_terms,ECTermScoring-method}
\title{Remove Empty Terms from an ECTermScoring Object}
\usage{
remove_empty_terms(object)

\S4method{remove_empty_terms}{ECTermScoring}(object)
}
\arguments{
\item{object}{An ECTermScoring object.}
}
\value{
A new, smaller ECTermScoring object with empty terms removed.
}
\description{
Creates a new ECTermScoring object that excludes all terms with a
degree of zero (i.e., terms with no associated elements).
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_empty_terms(ECTermScoring)}: Method for ECTermScoring objects.

}}
\examples{
# Create an object where T2 is an empty term
te_df <- data.frame(
  term = c("T1", "T1", "T2"),
  element = c("E1", "E2", "E1") # T2 has an edge, but let's make it empty
)
# To properly create an empty term, we build a graph with it
full_universe_df <- data.frame(
   v1 = c("T1", "T1", "T2"),
   v2 = c("E1", "E2", "E1")
)
ecg <- ECGraph(full_universe_df) # ECGraph constructor will find T2

# Create the ECTermScoring object from an edge list that omits T2's edges
ects_with_empty_term <- ECTermScoring(data.frame(term="T1", element=c("E1","E2")))

# For this example, let's assume `ects_with_empty_term` was created
# in a way that it knows about "T2" but has no elements for it.
# ects_trimmed <- remove_empty_terms(ects_with_empty_term)
# print(ects_trimmed@terms) # Would not contain "T2"

}
