% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{remove_isolated_elements}
\alias{remove_isolated_elements}
\alias{remove_isolated_elements,ECTermScoring-method}
\title{title Remove Isolated Elements from an ECTermScoring Object}
\usage{
remove_isolated_elements(object)

\S4method{remove_isolated_elements}{ECTermScoring}(object)
}
\arguments{
\item{object}{An ECTermScoring object.}
}
\value{
A new, smaller ECTermScoring object.
}
\description{
Creates a new ECTermScoring object that excludes all elements with
zero degree (no element). This method only removes isolated *elements*;
isolated *terms* (terms with no associated elements) are kept.
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_isolated_elements(ECTermScoring)}: Method for ECTermScoring objects.

}}
\examples{
# Create an object where E3 is an isolated element
te_df <- data.frame(
  term = c("T1", "T1"),
  element = c("E1", "E2")
)
# Manually create a universe that includes an extra element
full_universe_df <- data.frame(
   v1 = c("T1", "T1", "T2"), # T2 is an isolated term
   v2 = c("E1", "E2", "E3")  # E3 is an isolated element
)
ecg <- ECGraph(full_universe_df)
ects_base <- ECTermScoring(te_df) # E3 is not in this edge list

}
