% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{remove_isolated_vertices}
\alias{remove_isolated_vertices}
\alias{remove_isolated_vertices,ECGraph-method}
\title{Remove Isolated Vertices from an ECGraph}
\usage{
remove_isolated_vertices(object)

\S4method{remove_isolated_vertices}{ECGraph}(object)
}
\arguments{
\item{object}{An ECGraph object.}
}
\value{
A new, smaller ECGraph object with no isolated vertex.
}
\description{
Creates a new ECGraph object that excludes all vertices with
zero degree (isolated vertices).
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_isolated_vertices(ECGraph)}: Method for ECGraph objects.

}}
\examples{
# Create a graph with isolated vertex "D"
edge_df <- data.frame(p1 = c("A", "B", "D"), p2 = c("B", "C", "D"))
ecg <- ECGraph(edge_df)
ecg_trimmed <- remove_isolated_vertices(ecg)
print(ecg@names) # "A", "B", "C", "D"
print(ecg_trimmed@names) # "A", "B", "C"

}
