% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ecg}
\alias{sample_ecg}
\title{Sample Protein Interaction Network}
\format{
An object of class \code{\link{ECGraph}} with 7244 nodes (elements)
  and 286,341 edges.
}
\source{
\url{https://string-db.org/}
Data from STRING is licensed under CC BY 4.0:
\url{https://creativecommons.org/licenses/by/4.0/}
}
\usage{
sample_ecg
}
\description{
A curated subset of the human STRING network (version 12.0) with interaction
scores of 900 or higher. The network has been trimmed to create a smaller,
more manageable example for the package's vignettes and tests.
}
\keyword{datasets}
