% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_gene_symbol_lookup}
\alias{sample_gene_symbol_lookup}
\title{Entrez Gene Symbol Lookup Table}
\format{
A named character vector. The names are Entrez IDs, and the values are
  the gene symbols.
}
\source{
org.Hs.eg.db Bioconductor package
}
\usage{
sample_gene_symbol_lookup
}
\description{
A named character vector that maps Entrez Gene IDs (e.g., "7157") to their
corresponding official gene symbols (e.g., "TP53"). This is a helper dataset
for annotating results.
}
\keyword{datasets}
