% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_gene_symbols}
\alias{sample_gene_symbols}
\title{Sample Gene Symbols Lookup}
\format{
A named character vector.
}
\usage{
sample_gene_symbols
}
\description{
A named character vector mapping Entrez IDs to Gene Symbols.
}
\keyword{datasets}
