% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_term_names}
\alias{sample_term_names}
\title{Sample Term Names Lookup}
\format{
A named character vector.
}
\usage{
sample_term_names
}
\description{
A named character vector mapping Term IDs (GO IDs) to descriptive names.
}
\keyword{datasets}
