% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{summarize_suitability_fast}
\alias{summarize_suitability_fast}
\alias{summarize_suitability_fast,ECProb-method}
\title{Summarize Graph's Suitability for Fast Lambda Approximation}
\usage{
summarize_suitability_fast(object)

\S4method{summarize_suitability_fast}{ECProb}(object)
}
\arguments{
\item{object}{An ECProb object.}
}
\value{
A list of summary statistics: 'p_ij_over_1' = proportion of vertex pairs such
that p_ij > 1 and 'prop_problematic_vertices' = proportion of vertices with degree > sqrt(2*M)
}
\description{
Provides statistics on the pairwise Bernoulli parameters (p_ij)
to help assess if fast lambda approximation methods are suitable for the graph.
This method also reports the proportion of vertices that are problematic for fast
lambda approximation in the sense that their degree is greater than sqrt(2*M), M
being the graph size.
}
\section{Methods (by class)}{
\itemize{
\item \code{summarize_suitability_fast(ECProb)}: Method for ECProb objects

}}
