% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{terms_ecranks_statistics}
\alias{terms_ecranks_statistics}
\alias{terms_ecranks_statistics,ECTermScoring-method}
\title{Score Terms Against a Ranked List of Elements}
\usage{
terms_ecranks_statistics(object, element_ranks)

\S4method{terms_ecranks_statistics}{ECTermScoring}(object, element_ranks)
}
\arguments{
\item{object}{An ECTermScoring object.}

\item{element_ranks}{A named list or vector of element ranks.}
}
\value{
A named list where each element is a data frame of statistics for a term.
}
\description{
Calculates a running enrichment score for terms based on a ranked list of elements.
}
\details{
This method implements a fast, vectorized algorithm. For statistical
rigor, the analysis is conditioned on the universe of elements provided in the
`element_ranks` argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{terms_ecranks_statistics(ECTermScoring)}: Method for ECTermScoring objects.

}}
