% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{to_dataframe,ECTermScoring-method}
\alias{to_dataframe,ECTermScoring-method}
\title{Convert ECTermScoring to Data Frame}
\usage{
\S4method{to_dataframe}{ECTermScoring}(object)
}
\arguments{
\item{object}{An ECTermScoring object.}
}
\value{
A data frame with two columns "term" and "element" representing
the edge set of the bipartite graph.
}
\description{
Creates a data frame of edges from an ECTermScoring object.
The returned data frame has two columns representing term-element edges.
Isolated elements (no associated terms) and empty terms (no associated elements)
are not included in the output.
}
\examples{
# The to_dataframe generic is defined with the ECGraph class.
# This method provides the implementation for ECTermScoring objects.
te_df <- data.frame(
  term = c("T1", "T1", "T2"),
  element = c("E1", "E2", "E2")
)
ects <- ECTermScoring(te_df)
edge_list <- to_dataframe(ects)
print(edge_list)

}
