% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R, R/ECTermScoring.R
\name{to_dataframe}
\alias{to_dataframe}
\alias{to_dataframe,ECGraph-method}
\alias{to_dataframe,ECTermScoring-method}
\title{Convert ECGraph to Data Frame}
\usage{
to_dataframe(object)

\S4method{to_dataframe}{ECGraph}(object)

\S4method{to_dataframe}{ECTermScoring}(object)
}
\arguments{
\item{object}{An ECTermScoring object.}
}
\value{
A data frame with two columns "from" and "to" representing the edge set.

A data frame with two columns "term" and "element" representing
the edge set of the bipartite graph.
}
\description{
Creates a data frame of edges from an ECGraph object.
The returned data frame has two columns representing the undirected
edges contained in the ECGraph object. Isolated vertices are not included.

Creates a data frame of edges from an ECTermScoring object.
The returned data frame has two columns representing term-element edges.
Isolated elements (no associated terms) and empty terms (no associated elements)
are not included in the output.
}
\section{Methods (by class)}{
\itemize{
\item \code{to_dataframe(ECGraph)}: Method for ECGraph objects.

\item \code{to_dataframe(ECTermScoring)}: Method for ECTermScoring objects.

}}
\examples{
edge_df <- data.frame(vertex1 = c("A", "B"), vertex2 = c("B", "C"))
ecg <- ECGraph(edge_df)
edge_list_df <- to_dataframe(ecg)
print(edge_list_df)

# The to_dataframe generic is defined with the ECGraph class.
# This method provides the implementation for ECTermScoring objects.
te_df <- data.frame(
  term = c("T1", "T1", "T2"),
  element = c("E1", "E2", "E2")
)
ects <- ECTermScoring(te_df)
edge_list <- to_dataframe(ects)
print(edge_list)

}
