% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{trim_ecgraph}
\alias{trim_ecgraph}
\alias{trim_ecgraph,ECGraph-method}
\title{Trim High-Degree Vertices from an ECGraph}
\usage{
trim_ecgraph(object, threshold = 1)

\S4method{trim_ecgraph}{ECGraph}(object, threshold = 1)
}
\arguments{
\item{object}{An ECGraph object.}

\item{threshold}{A numeric threshold for the trimming condition. The default
value of 1 corresponds to the theoretical condition `k_max^2 >= 2M`. Lowering
this value will potentially result in better '_fast' approximations.}
}
\value{
A list containing two elements:
\itemize{
  \item `trimmed_graph`: A new, smaller ECGraph object with without high-degree vertices.
  \item `removed_vertices`: A character vector of the vertex names that were removed.
}
}
\description{
Removes the highest-degree vertices (hubs) from an
ECGraph object until the condition for fast lambda approximation
is met to the prescribed precision 'threshold' (no greater than 1).
}
\details{
This function provides a way to prepare a graph for analyses
that use the `_fast` lambda calculation methods. It works by:
1. Identifying the current maximum degree (`k_max`) in the graph.
2. Checking if `k_max^2 >= threshold * 2M`, where `M` is the current total number of edges.
3. If the condition is met, all vertices with degree `k_max` are removed.
4. The graph size `M` is updated, and the process repeats.
}
\section{Methods (by class)}{
\itemize{
\item \code{trim_ecgraph(ECGraph)}: Method for ECGraph objects.

}}
\examples{
# Create a star graph where 'A' is a hub
edge_df <- data.frame(
  from = "A",
  to = c("B", "C", "D", "E", "F")
)
ecg <- ECGraph(edge_df)

# The default condition is not met: k_max=5, M=5.  5*5 / (2*5) > 1
trimmed_result <- trim_ecgraph(ecg)

# The central hub 'A' was removed
print(trimmed_result$removed_vertices) # "A"

# The new graph contains only the isolated outer nodes
print(trimmed_result$trimmed_graph@adj)

}
