% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized,ECProb-method}
\title{Vectorized calculation of fast between-set statistics}
\usage{
calculate_between_stats_fast_vectorized(object, pairs_dt, set_membership_dt)

\S4method{calculate_between_stats_fast_vectorized}{ECProb}(object, pairs_dt, set_membership_dt)
}
\arguments{
\item{object}{An ECProb object.}

\item{pairs_dt}{A data table of pairs of element sets.}

\item{set_membership_dt}{A data table mapping set IDs to their elements.}
}
\value{
A final `data.table` with stats (`observed_edges`, `lambda`, etc.).
}
\description{
Calculates fast between-set statistics for a collection of set pairs
using a high-performance, vectorized algorithm.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_between_stats_fast_vectorized(ECProb)}: Method for ECProb objects

}}
\examples{
# Create sample data
ecg <- ECGraph(data.frame(p1=c("E1","E3"), p2=c("E4","E5")))
ecp <- ECProb(ecg)
pairs <- data.table(set1 = c("SA", "SC"), set2 = c("SB", "SD"))
memberships <- data.table(
  set_id = c("SA","SA","SB","SC","SC","SD"),
  element = c("E1","E2","E2","E3","E4","E5")
)
# Run vectorized calculation
calculate_between_stats_fast_vectorized(ecp, pairs, memberships)

}
