% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized,ECProb-method}
\title{Vectorized calculation of fast between statistics}
\usage{
calculate_between_stats_fast_vectorized(object, pairs_dt, set_membership_dt)

\S4method{calculate_between_stats_fast_vectorized}{ECProb}(object, pairs_dt, set_membership_dt)
}
\arguments{
\item{object}{An ECProb object.}

\item{pairs_dt}{A data table of pairs of element sets represented by pairs:
two columns "set1" and "set2" giving set IDs}

\item{set_membership_dt}{A data table mapping set IDs to their elements.
Must have "set_id" and "element" columns.}
}
\value{
A data.table with pairs_id, disjoint_set1, disjointt_set2 and list
 of statistics.
}
\description{
Calculates fast between statistics for a collection of set pairs
using vectorization. The method first transforms pairs to disjoint sets in a
vectorized way and then uses vectorization to calculate lambda, observed edge count
and associated statistics. Note that vectorization implies that the fast
approximation p_ij = k_ik_j/(2M) is utilized.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_between_stats_fast_vectorized(ECProb)}: Method for ECProb objects

}}
\examples{
TBD
}
