% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized}
\alias{calculate_between_stats_fast_vectorized,ECProb-method}
\title{Vectorized calculation of fast between statistics}
\usage{
calculate_between_stats_fast_vectorized(object, pairs_dt, set_membership_dt)

\S4method{calculate_between_stats_fast_vectorized}{ECProb}(object, pairs_dt, set_membership_dt)
}
\arguments{
\item{object}{An ECProb object.}

\item{pairs_dt}{A data table of pairs of element sets, with two columns
"set1" and "set2" giving the set IDs.}

\item{set_membership_dt}{A data table mapping set IDs to their elements.
Must have "set_id" and "element" columns.}
}
\value{
A final `data.table` with the original pairs and columns for all
  calculated statistics (`set1`, `size1`, `set2`, `size2`, `observed_edges`,
  `lambda`, `p_value`, `log2_Anscombe_ratio`).
}
\description{
Calculates fast between statistics for a collection of set pairs
using a high-performance, vectorized algorithm.
}
\details{
This function is designed for efficiency when processing many term
pairs at once. It uses a series of `data.table` joins and aggregations to
avoid slow R-level loops. Note that this method uses the "fast" lambda
approximation.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_between_stats_fast_vectorized(ECProb)}: Method for ECProb objects

}}
\examples{
# TBD
}
