% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_in_stats_fast_vectorized}
\alias{calculate_in_stats_fast_vectorized}
\alias{calculate_in_stats_fast_vectorized,ECProb-method}
\title{Vectorized calculation of fast within-set statistics}
\usage{
calculate_in_stats_fast_vectorized(object, sets_dt, set_membership_dt)

\S4method{calculate_in_stats_fast_vectorized}{ECProb}(object, sets_dt, set_membership_dt)
}
\arguments{
\item{object}{An ECProb object.}

\item{sets_dt}{A data table with a single column ("set_id") listing the
sets to be analyzed.}

\item{set_membership_dt}{A data table mapping set IDs to their elements.
Must have "set_id" and "element" columns.}
}
\value{
A final `data.table` with the input set IDs and columns for all
  calculated statistics (`observed_edges`, `lambda`, `p_value`, etc.).
}
\description{
Calculates fast within-set statistics for a collection of sets
using a high-performance, vectorized algorithm.
}
\details{
This function is designed for efficiency when processing many sets
at once. It uses a series of `data.table` joins and aggregations to
avoid slow R-level loops. Note that this method uses the "fast" lambda
approximation.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_in_stats_fast_vectorized(ECProb)}: Method for ECProb objects

}}
\examples{
# TBD
}
