% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_in_stats_fast_vectorized}
\alias{calculate_in_stats_fast_vectorized}
\alias{calculate_in_stats_fast_vectorized,ECProb-method}
\title{Vectorized calculation of fast within-set statistics}
\usage{
calculate_in_stats_fast_vectorized(object, sets_dt, set_membership_dt)

\S4method{calculate_in_stats_fast_vectorized}{ECProb}(object, sets_dt, set_membership_dt)
}
\arguments{
\item{object}{An ECProb object.}

\item{sets_dt}{A data table with a single column ("set_id")}

\item{set_membership_dt}{A data table with "set_id" and "element" columns.}
}
\value{
A data table with edge-count stattistics (`observed_edge_count`, `lambda`, etc.).
}
\description{
Calculates fast within-set statistics for a collection of sets
using vectorized algorithm.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_in_stats_fast_vectorized(ECProb)}: Method for ECProb objects

}}
\examples{
# Create sample data
ecg <- ECGraph(data.frame(p1=c("E1","E3"), p2=c("E2","E4")))
ecp <- ECProb(ecg)
sets <- data.table(set_id = c("SA", "SB"))
memberships <- data.table(
  set_id = c("SA","SA","SA","SB","SB"),
  element = c("E1","E2","E3","E3","E4")
)
# Run vectorized calculation
calculate_in_stats_fast_vectorized(ecp, sets, memberships)

}
