% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_lambda_between_fast}
\alias{calculate_lambda_between_fast}
\alias{calculate_lambda_between_fast,ECProb-method}
\title{Calculate Lambda Between Two Vertex Sets (Fast Method)}
\usage{
calculate_lambda_between_fast(object, set1, set2)

\S4method{calculate_lambda_between_fast}{ECProb}(object, set1, set2)
}
\arguments{
\item{object}{An ECProb object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{A character vector of vertex names.}
}
\value{
The approximate numeric lambda parameter.
}
\description{
Calculates an approximation of the lambda parameter between two
vertex sets.
}
\details{
This method is the fastest, with a complexity of O(n+m). It is an
approximation based on the assumption that the Bernoulli probability of an edge
between any two vertices i and j, p_ij = (k_i * k_j) / (2M), is always less than 1.
The accuracy of this method decreases as the proportion of vertex pairs for which
p_ij > 1 (let's call this proportion alpha) increases. You can estimate alpha
for your graph using \code{\link{summarize_suitability_fast}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_lambda_between_fast(ECProb)}: Method for ECProb objects.

}}
\examples{
ecg <- ECGraph(data.frame(p1=c("A","B","C"), p2=c("X","Y","Z")))
ecp <- ECProb(ecg)
calculate_lambda_between_fast(ecp, c("A", "B"), c("X", "Y"))

}
