% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{calculate_p_value}
\alias{calculate_p_value}
\alias{calculate_p_value,ECProb-method}
\title{Calculate P-value for an Observed Edge Count}
\usage{
calculate_p_value(object, z, m, lambda)

\S4method{calculate_p_value}{ECProb}(object, z, m, lambda)
}
\arguments{
\item{object}{An ECProb object (for S4 dispatch).}

\item{z}{The observed number of edges.}

\item{m}{The maximum possible number of edges.}

\item{lambda}{The Poisson parameter (expected edge count).}
}
\value{
The calculated p-value.
}
\description{
Calculates the p-value for an observed edge count based on a
truncated Poisson distribution.
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_p_value(ECProb)}: Method for ECProb objects.

}}
\examples{
ecg <- ECGraph(data.frame(p1="A", p2="B")) # Dummy graph
ecp <- ECProb(ecg)
# P-value for observing 5 edges when max is 10 and lambda is 2
calculate_p_value(ecp, z = 5, m = 10, lambda = 2)

}
