% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{ECGraph}
\alias{ECGraph}
\title{Constructor for ECGraph Objects}
\usage{
ECGraph(edges_input, col1 = 1, col2 = 2, header = TRUE, sep = "\\t")
}
\arguments{
\item{edges_input}{A character string specifying the path to a file,
a data frame with two columns representing edges, or an \code{igraph} graph object.
Vertex IDs will be coerced to character.}

\item{col1}{The name or index of the first column if `edges_input` is a
data frame or file. Defaults to 1.}

\item{col2}{The name or index of the second column if `edges_input` is a
data frame or file. Defaults to 2.}

\item{header}{A logical value for reading from a file. Defaults to TRUE.}

\item{sep}{The field separator character for reading from a file. Defaults to "\\t".}
}
\value{
An object of class \code{\linkS4class{ECGraph}}.
}
\description{
Creates an \code{\linkS4class{ECGraph}} object from a file, data frame, or igraph object.
}
\examples{
# --- Data ---
edge_df <- data.frame(
   vertex1 = c("A", "A", "B", "C", "E"),
   vertex2 = c("B", "C", "C", "D", "E"),
   stringsAsFactors = FALSE
)
temp_file <- tempfile(fileext = ".tsv")
write.table(edge_df, temp_file, sep = "\t", row.names = FALSE, quote = FALSE)

# --- ECGraph from a data frame ---
graph_from_df <- ECGraph(edge_df)

# --- ECGraph from a file path ---
graph_from_file <- ECGraph(temp_file)

unlink(temp_file)
}
\seealso{
Functions that operate on ECGraph objects:
\itemize{
   \item \code{\link{to_dataframe}}: Convert an ECGraph to a data frame edge list.
   \item \code{\link{get_neighbors}}: Retrieve all neighbors for a set of vertices.
   \item \code{\link{get_edge_count_in}}: Count edges within a set of vertices.
   \item \code{\link{get_edge_count_between}}: Count edges between two sets of vertices.
   \item \code{\link{get_edge_count_in_max}}: Maximum possible edge count within a set (complete graph).
   \item \code{\link{get_edge_count_between_max}}: Maximum possible edge count between two sets (complete graph).
   \item \code{\link{get_edge_count_in_max_fds}}: Maximum possible edge count within a set (fixed degree sequence).
   \item \code{\link{get_edge_count_between_max_fds}}: Maximum possible edge count between two sets (fixed degree sequence).
   \item \code{\link{trim_ecgraph}}: Trim high-degree vertices (hubs).
   \item \code{\link{remove_isolated_vertices}}: Remove isolated vertices.
   \item \code{\link{show}}: Summary display of the object.
}
}
