% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\docType{class}
\name{ECProb}
\alias{ECProb}
\alias{ECProb-class}
\title{ECProb S4 Class and Constructor}
\arguments{
\item{ecgraph}{An ECGraph object.}
}
\value{
An object of class ECProb.
}
\description{
Extends the ECGraph class to include pre-calculated properties
necessary for edge-count probability calculations.
}
\section{Slots}{

\describe{
\item{\code{graph_size}}{numeric. The total number of edges in the graph (M).}

\item{\code{graph_order}}{numeric. The total number of vertices in the graph (N).}

\item{\code{average_degree}}{numeric. The average degree of the vertices in the graph.}

\item{\code{adj}}{list. (Inherited from ECGraph) The adjacency list.}

\item{\code{degrees}}{list. (Inherited from ECGraph) The degree of each vertex.}

\item{\code{names}}{character. (Inherited from ECGraph) The names of all vertices.}
}}

\examples{
edge_df <- data.frame(p1 = c("A", "B", "C"), p2 = c("B", "C", "D"))
ecg <- ECGraph(edge_df)
ecp <- ECProb(ecg)
ecp@graph_size # 3

}
\seealso{
The base class for graph representation: \code{\link{ECGraph}}
}
