% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\docType{class}
\name{ECTermScoring-class}
\alias{ECTermScoring-class}
\alias{show,ECTermScoring-method}
\title{ECTermScoring S4 Class}
\usage{
\S4method{show}{ECTermScoring}(object)
}
\arguments{
\item{object}{An ECTermScoring object}
}
\description{
The main class for EdgeCount analysis with bipartite graphs, representing a bipartite
network of term-element memberships. It extends the ECProb class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(ECTermScoring)}: Show method for ECTermScoring objects.

}}
\section{Slots}{

\describe{
\item{\code{ecprob}}{An object of class ECProb, representing the underlying
undirected version of the bipartite graph.}

\item{\code{elements}}{A character vector of all unique vertex IDs identified as elements.}

\item{\code{terms}}{A character vector of all unique vertex IDs identified as terms.}
}}

\seealso{
\code{\link{ECTermScoring}} for the constructor.
}
