% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{edge_count_statistics}
\alias{edge_count_statistics}
\alias{edge_count_statistics,ECProb-method}
\title{Calculate Full Edge Count Statistics}
\usage{
edge_count_statistics(
  object,
  set1,
  set2 = NULL,
  observed_edge_count,
  lambda_method = "optimized"
)

\S4method{edge_count_statistics}{ECProb}(
  object,
  set1,
  set2 = NULL,
  observed_edge_count,
  lambda_method = "optimized"
)
}
\arguments{
\item{object}{An ECProb object.}

\item{set1}{A character vector of vertex names.}

\item{set2}{(Optional) A second character vector for between-set analysis.}

\item{observed_edge_count}{The observed number of edges.}

\item{lambda_method}{The method for lambda calculation ("accurate", "optimized", "fast").}
}
\value{
A list containing the p-value, lambda, log2_Anscombe_ratio and log2_relative_change.
}
\description{
A wrapper that calculates a suite of statistics
for an observed edge count.
}
\section{Methods (by class)}{
\itemize{
\item \code{edge_count_statistics(ECProb)}: Method for ECProb objects.

}}
