% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{get_disjoint_sets}
\alias{get_disjoint_sets}
\title{Get Disjoint Sets for Multiple Pairs (Vectorized)}
\usage{
get_disjoint_sets(pairs_dt, set_membership_dt)
}
\arguments{
\item{pairs_dt}{A `data.table` with two columns ("set1", "set2").}

\item{set_membership_dt}{A `data.table` with "set_id" and "element" columns.}
}
\value{
A `data.table` with "set1", "set2", "elements1_disjoint" (list),
  and "elements2_disjoint" (list).
}
\description{
An internal function that calculates the
disjoint element sets for a list of set pairs.
}
