% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{reduce_universe_by_elements}
\alias{reduce_universe_by_elements}
\alias{reduce_universe_by_elements,ECTermScoring-method}
\title{Reduce the Universe of an ECTermScoring Object to a Set of Elements}
\usage{
reduce_universe_by_elements(object, elements_to_keep)

\S4method{reduce_universe_by_elements}{ECTermScoring}(object, elements_to_keep)
}
\arguments{
\item{object}{An ECTermScoring object.}

\item{elements_to_keep}{A character vector of element IDs to keep.}
}
\value{
A new, smaller ECTermScoring object.
}
\description{
Creates a new, smaller ECTermScoring object by restricting it
to a specified subset of elements.
}
\details{
This is a helper for ensuring statistical tests are correctly
conditioned. It takes an ECTermScoring object and a vector of elements, finds
the intersection, and returns a new object where all properties (degrees, graph
size, etc.) are recalculated based on this smaller universe. This method is an
internal helper for rank scoring.
}
\section{Methods (by class)}{
\itemize{
\item \code{reduce_universe_by_elements(ECTermScoring)}: Method for ECTermScoring objects.

}}
