% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{run_vsea}
\alias{run_vsea}
\alias{run_vsea,ECTermScoring-method}
\title{Perform Vertex Set Enrichment Analysis}
\usage{
run_vsea(
  object,
  element_ranks,
  scoring_statistic = "log2_Anscombe_ratio",
  n_permutations = 1000,
  seed = NULL
)

\S4method{run_vsea}{ECTermScoring}(
  object,
  element_ranks,
  scoring_statistic = "log2_Anscombe_ratio",
  n_permutations = 1000,
  seed = NULL
)
}
\arguments{
\item{object}{An ECTermScoring object.}

\item{element_ranks}{A named list or vector of element ranks.}

\item{scoring_statistic}{The score to use for enrichment (e.g., "log2_Anscombe_ratio").}

\item{n_permutations}{The number of permutations to generate the null
distribution, 1000 by default.}

\item{seed}{An optional random seed for reproducibility.}
}
\value{
A list containing a single data frame: \code{max_score_summary}.
  The data frame is sorted by \code{fdr_q_value} and contains:
  \itemize{
    \item \code{term_id}: The term identifier.
    \item \code{max_score}: The maximum running enrichment score observed.
    \item \code{nes}: Normalized Enrichment Score.
    \item \code{fdr_q_value}: Empirical False Discovery Rate.
    \item \code{term_size}: Number of elements in the term.
    \item \code{rank_at_max}: The element rank at which the score peaked.
    \item \code{observed_ec}: The number of observed edges at the peak.
  }
}
\description{
Runs a ranked-list enrichment analysis (Vertex Set Enrichment Analysis),
calculating a Normalized Enrichment Score (NES) and a False Discovery Rate (FDR) q-value
for each term using a permutation-based null distribution.
}
\details{
This version focuses solely on identifying positive enrichment,
i.e., the maximum score for each term
}
\section{Methods (by class)}{
\itemize{
\item \code{run_vsea(ECTermScoring)}: Method for ECTermScoring objects.

}}
\examples{
library(EdgeCount)
library(data.table)

# Load data
data(sample_ects)

# Random uniform ranking
element_ranks <- setNames(
  sample(seq_along(sample_ects@elements)),
  sample_ects@elements
)

# VSEA analysis
vsea_results <- run_vsea(
  sample_ects,
  element_ranks,
  n_permutations = 10, # Low for speed
  seed = 1
)

print(head(vsea_results$max_score_summary))
}
