% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECTermScoring.R
\name{summarize_ranks_full}
\alias{summarize_ranks_full}
\title{Full Summary of Rank Statistics (}
\usage{
summarize_ranks_full(
  term_scores_list,
  scoring_statistic = "log2_Anscombe_ratio"
)
}
\arguments{
\item{term_scores_list}{A named list of data.tables from `terms_ecranks_statistics`.}

\item{scoring_statistic}{The column name of the score to summarize.}
}
\value{
A `data.table` with one row per term, containing the rich summary.
}
\description{
A function that takes the output of
`terms_ecranks_statistics` and calculates a summary (min, max, median
scores and their context) for each term.
}
