% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECProb.R
\name{summarize_suitability_fast}
\alias{summarize_suitability_fast}
\alias{summarize_suitability_fast,ECProb-method}
\title{Summarize Graph's Suitability for Fast Lambda Approximation}
\usage{
summarize_suitability_fast(object)

\S4method{summarize_suitability_fast}{ECProb}(object)
}
\arguments{
\item{object}{An ECProb object.}
}
\value{
A list of summary statistics for the p_ij distribution.
}
\description{
Provides statistics on the pairwise Bernoulli parameters (p_ij)
to help assess if fast lambda approximation methods are suitable for the graph.
}
\section{Methods (by class)}{
\itemize{
\item \code{summarize_suitability_fast(ECProb)}: Method for ECProb objects

}}
