% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECGraph.R
\name{to_dataframe}
\alias{to_dataframe}
\alias{to_dataframe,ECGraph-method}
\title{Convert ECGraph to Data Frame}
\usage{
to_dataframe(object)

\S4method{to_dataframe}{ECGraph}(object)
}
\arguments{
\item{object}{An ECGraph object.}
}
\value{
A data frame with two columns "from" and "to" representing the edge set.
}
\description{
Creates a data frame of edges from an ECGraph object.
The returned data frame has two columns representing the undirected
edges contained in the ECGraph object. Isolated vertices are not included.
}
\section{Methods (by class)}{
\itemize{
\item \code{to_dataframe(ECGraph)}: Method for ECGraph objects.

}}
\examples{
edge_df <- data.frame(vertex1 = c("A", "B"), vertex2 = c("B", "C"))
ecg <- ECGraph(edge_df)
edge_list_df <- to_dataframe(ecg)
print(edge_list_df)

}
